% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.legend.R
\name{add.legend}
\alias{add.legend}
\title{Add a LaTeX-typeset legend to a plotFun or plotPoints plot}
\usage{
add.legend(
  ...,
  space = "right",
  columns = NULL,
  background = "white",
  default_lty = 1,
  default_lwd = 1,
  default_pch = NA,
  plot = lattice::trellis.last.object()
)
}
\arguments{
\item{...}{Legend entries.}

\item{space}{Legend position. Common values are `"right"`, `"top"`,
`"bottom"`, and `"left"`. Defaults to `"right"`.}

\item{columns}{Number of legend columns. If `NULL`, the default is `1`
for legends on the left or right, and the number of legend entries for
legends on the top or bottom.}

\item{background}{Background color for the legend. Defaults to `"white"`.}

\item{default_lty}{Default line type. Defaults to `1`.}

\item{default_lwd}{Default line width. Defaults to `1`.}

\item{default_pch}{Default plotting character. Defaults to `NA`, meaning no point.}

\item{plot}{A trellis plot object. Defaults to the most recent trellis plot.}
}
\value{
The updated trellis plot object.
}
\description{
Adds a lattice-style legend/key to a trellis plot, such as the output of
`plotFun()` or `plotPoints()`. This function is intended to be called once
after the plot has been built. If the plot already has a legend, it is
removed and replaced by the new legend.
}
\details{
Legend labels are processed with `latex2exp::TeX()`.

Legend entries are passed through `...`. The name of each entry is the
legend label. Each entry may be either a color string or a named vector
with entries `col`, `lty`, `lwd`, and/or `pch`.
}
\examples{
p <- plotFun(x^2 ~ x, xlim = c(-3, 3), col = "red", lty = 2, lwd = 3)
p <- plotFun(x^3 ~ x, add = TRUE, plot = p, col = "blue")

add.legend(
  "$y=x^2$" = c(col = "red", lty = 2, lwd = 3),
  "$y=x^3$" = "blue",
  plot = p
)

add.legend("$y$" = "maroon")

}
