% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalPlotting.R
\name{blank.canvas}
\alias{blank.canvas}
\title{Create a blank canvas with convenient limits for later use.}
\usage{
blank.canvas(xlim, ylim, xlab = "x", ylab = "y", ...)
}
\arguments{
\item{xlim}{the x limits}

\item{ylim}{the y limits}

\item{xlab}{the label to put on the x-axis}

\item{ylab}{the label to put on the y-axis.}

\item{...}{additional arguments to pass to lattice::xyplot of particular interest might be asp=1 for setting the aspect ratio of the plot.
blank.canvas(xlim=c(0,1),ylim=c(0,1),xlab="This is an axis", ylab="This too")
blank.canvas(xlim=c(0,1),ylim=c(0,1),xlab="This is an axis", ylab="This too",asp=1)}
}
\description{
Create a blank canvas with convenient limits for later use.
}
