% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Explosion1_doc.R
\docType{data}
\name{Explosion1}
\alias{Explosion1}
\title{Blast data for nuclear explosion test in White Sands, New Mexico}
\format{
A data frame with 23 observations and 2 variables:
\describe{
  \item{Time}{in seconds}
  \item{Radius}{radius of blast in meters}
}
}
\source{
Taylor, G. “The Formation of a Blast Wave by a Very Intense Explosion. II. The Atomic Explosion of 1945.” Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences, 1950, 201: 175–86.
}
\usage{
Explosion1
}
\description{
A modified dataset containing the blast radius (in meters) as a function of time (in seconds) for a nuclear test conducted by the United States in White Sands, New Mexico in 1945.
}
\keyword{datasets}
