% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalPlotting.R
\name{mathaxis.on}
\alias{mathaxis.on}
\title{Plots "traditional" x-y axes.}
\usage{
mathaxis.on(
  lwd = 3,
  col = "black",
  xat = 0,
  yat = 0,
  xextent = 1e+30,
  yextent = 1e+30,
  ...
)
}
\arguments{
\item{lwd}{the line width of the axes}

\item{col}{the color of the axes}

\item{xat}{the y-coordinate where the axis will be drawn.}

\item{yat}{the x-coordinate where the axis will be drawn.}

\item{xextent}{the x-range over which the x-axis will be plotted. If it isn't big enough by default, make it bigger.}

\item{yextent}{the y-range over which the x-axis will be plotted. If it isn't big enough by default, make it bigger.}
}
\description{
Plots "traditional" x-y axes.
}
\examples{
f=mosaic::makeFun(sin(x)~x)
plotFunFill(f(x)~x,xlim=c(-3,3))
mathaxis.on()
}
