% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalPlotting.R
\name{place.text}
\alias{place.text}
\title{Places text on a figure.  Supports some latex symbols, but backslashes must be double backslashes.}
\usage{
place.text(
  text,
  x,
  y,
  col = "black",
  zoom = 1,
  plot = lattice::trellis.last.object(),
  tex = TRUE,
  ...
)
}
\arguments{
\item{text}{The message to display}

\item{x}{The x-coordinate}

\item{y}{The y-coordinate}

\item{col}{the text color}

\item{zoom}{the factor by which to expand the text.}

\item{plot}{the figure to add the text to.}

\item{tex}{if TRUE, try to interpret this string as TeX code.}

\item{...}{additional options that are passed to the base `text` command. `?text` for options.}
}
\description{
Places text on a figure.  Supports some latex symbols, but backslashes must be double backslashes.
}
\examples{
plotFunFill(x^2~x)
place.text("x^2",0.5,0.8)

plotFun(pnorm(x)~x,xlim=c(-3,3),lwd=2)
mathaxis.on()
plotFunFill(pnorm(x)~x,xlim=c(-3,1.1),col='magenta',add=TRUE)
place.text("$x$",1.1,-0.04,cex=1.3)
place.text("$\\\\int_{-\\\\infty}^x \\\\frac{1}{\\\\sqrt{2\\\\pi}} e^{-\\\\frac{x^2}{2}}\\\\,dx$",0.5,0.3,cex=1.4)
}
