% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunWork.R
\name{plotFun}
\alias{plotFun}
\title{Plot a function of one or two variables}
\usage{
plotFun(
  object,
  ...,
  plot = lattice::trellis.last.object(),
  add = NULL,
  under = FALSE,
  xlim = NULL,
  ylim = NULL,
  npts = NULL,
  ylab = NULL,
  xlab = NULL,
  zlab = NULL,
  filled = TRUE,
  levels = NULL,
  nlevels = 10,
  labels = TRUE,
  surface = FALSE,
  groups = NULL,
  col = NULL,
  col.regions = topo.colors,
  type = "l",
  lwd = 2,
  lty = lattice::trellis.par.get("superpose.line")$lty,
  alpha = NULL,
  discontinuities = NULL,
  discontinuity = 1,
  showcolorscale = FALSE,
  interactive = mosaic::rstudio_is_available()
)
}
\arguments{
\item{object}{a mathematical expression or a function "of one variable" which will
converted to something intuitively equivalent to \code{object(x) ~ x}. (See examples)}

\item{...}{additional parameters, typically processed by \code{lattice} functions such as
\code{\link[lattice:xyplot]{lattice::xyplot()}}, \code{\link[lattice:levelplot]{lattice::levelplot()}} or their panel functions.
Frequently used parameters include
\describe{
\item{\code{main}}{main title for plot }
\item{\code{sub}}{subtitle for plot }
\item{\code{lwd}}{line width }
\item{\code{lty}}{line type }
\item{\code{col}}{a color or a (small) integer indicating which color in the current
color scheme is desired.}
}
Additionally, these arguments can be used to specify parameters for the function being
plotted and to specify the plotting window with natural names.  See the examples for such usage.}

\item{plot}{a trellis object; by default, the most recently created trellis plot.
When \code{add} is \code{TRUE}, the new function will
be plotted into a layer added to this object.}

\item{add}{if \code{TRUE}, then add a layer to an existing plot rather than creating a new plot.
If \code{NULL}, this will be determined by the value of \code{under}.}

\item{under}{if \code{TRUE}, then new layer is added beneath existing layers}

\item{xlim}{limits for x axis (or use variable names, see examples)}

\item{ylim}{limits for y axis (or use variable names, see examples)}

\item{npts}{number of points for plotting.}

\item{ylab}{label for y axis}

\item{xlab}{label for x axis}

\item{zlab}{label for z axis (when in surface-plot mode)}

\item{filled}{fill with color between the contours (\code{TRUE} by default)}

\item{levels}{levels at which to draw contours}

\item{nlevels}{number of contours to draw (if \code{levels} not specified)}

\item{labels}{if \code{FALSE}, don't label contours}

\item{surface}{draw a surface plot rather than a contour plot}

\item{groups}{grouping argument ala lattice graphics}

\item{col}{vector of colors for line graphs and contours}

\item{col.regions}{a vector of colors or a function (\code{topo.colors} by default) for generating such}

\item{type}{type of plot (\code{"l"} by default)}

\item{lwd}{vector of line widths for line graphs}

\item{lty}{vector of line types for line graphs}

\item{alpha}{number from 0 (transparent) to 1 (opaque) for the fill colors}

\item{discontinuities}{a vector of input values at which a function is
discontinuous or \code{NULL} to use
a heuristic to auto-detect.}

\item{discontinuity}{a positive number determining how sensitive the plot is to
potential discontinuity.  Larger values result in less sensitivity.  The default is 1.
Use \code{discontinuity = Inf} to disable discontinuity detection.  Discontinuity detection
uses a crude numerical heuristic and may not give the desired results in all cases.}

\item{interactive}{a logical indicating whether the surface plot should be
interactive.}
}
\description{
USAFACalc::plotFun is just a few bugfixes on mosaic::plotFun.  It also defaults npts to
200 for contour plots.  The bug fixes allow you to set line color, width, and type on contour plots.
}
