% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVectorField.R
\name{plotPhasePlane}
\alias{plotPhasePlane}
\title{Plot the phase plane for a vector field.}
\usage{
plotPhasePlane(
  ddt,
  xlim = c(-5, 5),
  ylim = c(-5, 5),
  add = FALSE,
  N = 20,
  col = "cornflowerblue",
  lwd = 2,
  plot = lattice::trellis.last.object(),
  ...
)
}
\arguments{
\item{ddt}{A vector-valued function that defines the RHS of your system of ODEs.}

\item{xlim}{the limits of the horizontal axis.  Can also use a name that matches your variables.}

\item{ylim}{the limits of the vertical axis. Can also use a name that matches your variables.}

\item{add}{T/F variable indicating whether this should be added to plot}

\item{N}{the number of horizontal and vertical rows of arrows to draw}

\item{col}{arrow color}

\item{lwd}{arrow line width}

\item{plot}{the plot to which this should be added, default is the current plot.}

\item{...}{additional arguments to pass to xyplot arrows.}
}
\description{
Plot the phase plane for a vector field.
}
\examples{
# example code
f=mosaic::makeFun(c(-y,x)~x&y)
plotPhasePlane(f(x,y)~x&y, xlim=c(-3,3),ylim=c(-4,4))

#Classic predator-prey.
ode=mosaic::makeFun(c(2*rabbits-0.5*rabbits*foxes, -1*foxes+0.25*rabbits*foxes)~rabbits&foxes)
plotPhasePlane(ode(rabbits,foxes)~rabbits&foxes,rabbitslim=c(0,10),foxeslim=c(0,10))
mosaic::plotPoints(c(0,4)~c(0,4),pch=19,cex=1.3,col="magenta",add=TRUE)
soln=USAFACalc::Euler(ode(rabbits,foxes)~rabbits&foxes,ic=c(4,2),tlim=c(0,10),stepSize=0.001)
mosaic::plotPoints(foxes~rabbits,data=soln,pch=".",cex=1.3,add=TRUE,col="forestgreen")
}
