% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVectorField.R
\name{plotVectorField}
\alias{plotVectorField}
\title{Plot a vector field
Scales all vectors to have same length}
\usage{
plotVectorField(
  expression,
  xlim = c(-5, 5),
  ylim = c(-5, 5),
  N = 20,
  col = "cornflowerblue",
  lwd = 2,
  normalize = FALSE,
  add = FALSE,
  xlist = NA,
  ylist = NA,
  plot = lattice::trellis.last.object(),
  ...
)
}
\arguments{
\item{expression}{A function of two variables that returns two ouputs to plot.}

\item{xlim}{the limits of the horizontal axis}

\item{ylim}{the limits of the vertical axis}

\item{N}{the number of horizontal and vertical rows of arrows to draw}

\item{col}{arrow color}

\item{lwd}{arrow line width}

\item{normalize}{if true, then plot all vectors as unit length}

\item{add}{T/F variable indicating whether this should be added to plot}

\item{xlist}{the horizontal grid points to use in the plot, if specified.}

\item{ylist}{the vertical grid points to use in the plot, if specified.}

\item{plot}{the plot to which this should be added, default is the current plot.}

\item{...}{additional arguments to pass to xyplot arrows.}
}
\description{
Plot a vector field
Scales all vectors to have same length
}
\examples{
# example code
f=mosaic::makeFun(c(-y,x)~x&y)
plotVectorField(f(x,y)~x&y, xlim=c(-3,3),ylim=c(-4,4))

#You can set the color, line width, or line type as well.
plotVectorField(f(x,y)~x&y, xlim=c(-3,3),ylim=c(-4,4),col="green",lwd=3)

#You can add it to other plots if you so choose.
plotFun(sin(x)~x,xlim=c(0,5))
plotVectorField(f(x,y)~x&y, xlim=c(-3,3),ylim=c(-4,4),col="green",lwd=3,add=TRUE)
}
