% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwt_africa_doc.R
\docType{data}
\name{pwt_africa}
\alias{pwt_africa}
\title{GDP Data for Africa}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{country}{Country Name}
  \item{isocode}{ISO 3166-1 alpha-3 country code.}
  \item{year}{Year}
  \item{rgdpo}{Output-side real GDP at chained PPPs (in million 2017 USD)}
  \item{pop}{Population (in millions)}
  \item{rnna}{Capital stock at constant 2017 national prices (in million 2017 USD)}
  \item{region}{region the country is in}
  ...
}
}
\usage{
pwt_africa
}
\description{
GDP Data for Africa
}
\keyword{datasets}
