% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReimannVisualization.R
\name{Reimann.MP}
\alias{Reimann.MP}
\title{Visualize the midpoint rule}
\usage{
Reimann.MP(f, xlim = c(0, 5), N = 10)
}
\arguments{
\item{f}{an expression for the function}

\item{xlim}{limits of integration}

\item{N}{number of rectangles.}
}
\description{
Visualize the midpoint rule
}
