% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReimannVisualization.R
\name{Riemann.LH}
\alias{Riemann.LH}
\title{Visualize the left-hand rule}
\usage{
Riemann.LH(f, xlim = c(0, 5), N = 10)
}
\arguments{
\item{f}{an expression for the function}

\item{xlim}{limits of integration}

\item{N}{number of rectangles.}
}
\description{
Visualize the left-hand rule
}
\examples{
Riemann.LH(exp(-x^2/2)~x,xlim=c(-5,5),N=10)

}
