% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T53ClimbData2_doc.R
\docType{data}
\name{T53ClimbData2}
\alias{T53ClimbData2}
\title{Fictional T53 Climb Dataset}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{Altitute}{Altitute, in feet}
  \item{Time}{Elapsed time in seconds. }
  ...
}
}
\usage{
T53ClimbData2
}
\description{
Contains fictional, but realistic, data on the climb rate of a T53.
}
\keyword{datasets}
