% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreeBladeProp_T53_doc.R
\docType{data}
\name{ThreeBladeProp_T53}
\alias{ThreeBladeProp_T53}
\title{Data for a three-bladed T53}
\format{
A data frame with 12 observations and 2 variables:
\describe{
  \item{Time}{in seconds}
  \item{Altitude}{in feet above the ground.}
}
}
\usage{
ThreeBladeProp_T53
}
\description{
Contains fictional, but realistic data on the climb performance of a T-53 aircraft outfitted with a three bladed propeller.
}
\keyword{datasets}
