% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TurboNormalizerT53_doc.R
\docType{data}
\name{TurboNormalizer_T53}
\alias{TurboNormalizer_T53}
\title{Data for a T53 with a turbonormalizer}
\format{
A data frame with 13 observations and 2 variables:
\describe{
  \item{Time}{in seconds}
  \item{Altitude}{in feet above the ground.}
}
}
\usage{
TurboNormalizer_T53
}
\description{
Contains fictional, but realistic data on the climb performance of a T-53 aircraft outfitted with a turbonormalizing system.
}
\keyword{datasets}
