% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateUSAFACalc.R
\name{updateUSAFACalc}
\alias{updateUSAFACalc}
\title{Updates the USAFACalc package, if update is available, from github. Does not
seek to update dependencies.}
\usage{
updateUSAFACalc()
}
\description{
Updates the USAFACalc package, if update is available, from github. Does not
seek to update dependencies.
}
