% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USFoodExpenditure_doc.R
\docType{data}
\name{USFoodExpenditure}
\alias{USFoodExpenditure}
\title{US Food Expenditures}
\format{
A data frame with 61 observations and 2 variables:
\describe{
  \item{Year}{The year of observation}
  \item{Percent}{Percent of disposable income spend on food.}
  ...
}
}
\source{
https://humanprogress.org/. Accessed on July 15, 2025.
}
\usage{
USFoodExpenditure
}
\description{
A dataset containing total food expenditure of American families as a percentage of disposable personal income as a function of the year from 1950 to 2010.
}
\keyword{datasets}
