% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerMethod.R
\name{Euler}
\alias{Euler}
\title{Run Euler's method on an ODE.}
\usage{
Euler(dydt, tlim, y0, stepSize = (tlim[[2]] - tlim[[1]])/10)
}
\arguments{
\item{dydt}{an expression of t and y, in that order,  defining the right hand side of the ODE.}

\item{tlim}{a list containing the start and end of the timeframe to integrate over}

\item{y0}{the initial condition}

\item{stepSize}{length of timestep to take.}
}
\value{
A data frame containing the approximations of the solutions and the corresponding times.
}
\description{
Run Euler's method on an ODE.
}
\examples{
soln=Euler(y*(y-1) ~ t&y, tlim=c(0,10),y0=0.99)
soln

results=Euler(5*y*(y-1)~t&y,tlim=c(0,10),y0=0.99,stepSize=0.1)
results
mosaic::plotPoints(y~t,data=results)
}
