% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalPlotting.R
\name{grid.on}
\alias{grid.on}
\title{Turns on a grid for easier viewing.}
\usage{
grid.on(
  h = -1,
  v = -1,
  lwd = 2,
  col = "black",
  lty = 2,
  plot = lattice::trellis.last.object()
)
}
\arguments{
\item{h}{# of horizontal grid lines to make, with -1 meaning to just line up to tickmarks.}

\item{v}{# of vertical grid lines to make, with -1 meaning to just line up to tickmarks.}

\item{lwd}{line width, 2 is about right.}

\item{col}{grid color}

\item{lty}{line type, 1,2,or 3 are good options.}

\item{plot}{the figure to which to add the grid.
#@examples
plotFunFill(x^2~x)
grid.on()}
}
\description{
Turns on a grid for easier viewing.
}
