% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFunctionReplacements.R
\name{integrate}
\alias{integrate}
\title{integrate an expression over a finite interval.  A wrapper for stats::integrate}
\usage{
integrate(expression, xlim = NA, ...)
}
\arguments{
\item{expression}{the expression to integrate}

\item{xlim}{the interval, as a list, over which to integrate.}

\item{...}{intead of xlim, you may provide a limit that reflects the variable name.  For instance, if the variable is t, then tlim=c(-1,1)}
}
\value{
The definite integral requested.
}
\description{
integrate an expression over a finite interval.  A wrapper for stats::integrate
}
\examples{
integrate(x^2~x, xlim=c(-4,4))
#Integrate returns just a plain number now, so you can do stuff like:
int=integrate(x^2~x,xlim=c(-4,4))
int
integrate(t^2~t,tlim=c(0,5))

}
