% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalPlotting.R
\name{mathaxis.on}
\alias{mathaxis.on}
\title{Plots "traditional" x-y axes.}
\usage{
mathaxis.on(
  lwd = 3,
  col = "black",
  xat = 0,
  yat = 0,
  plot = lattice::trellis.last.object(),
  ...
)
}
\arguments{
\item{lwd}{the line width of the axes}

\item{col}{the color of the axes}

\item{xat}{the y-coordinate where the axis will be drawn.}

\item{yat}{the x-coordinate where the axis will be drawn.}

\item{plot}{the figure to which to add axes.}

\item{...}{additional options that will be passed to}
}
\description{
Plots "traditional" x-y axes.
}
\examples{
f=mosaic::makeFun(sin(x)~x)
plotFunFill(f(x)~x,xlim=c(-3,3))
mathaxis.on()
}
