% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalPlotting.R
\name{place.text}
\alias{place.text}
\title{Places text on a figure.}
\usage{
place.text(text, x, y, col = "black", ...)
}
\arguments{
\item{text}{The message to display}

\item{x}{The x-coordinate}

\item{y}{The y-coordinate}

\item{col}{the text color}
}
\description{
Places text on a figure.
}
\examples{
plotFunFill(x^2~x)
place.text("x^2",1,0.8)

}
