% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additionalPlotting.R
\name{plotFunFill}
\alias{plotFunFill}
\title{Plots a filled region defined by one or two functions.}
\usage{
plotFunFill(
  f,
  bottom = 0,
  xlim = c(0, 1),
  col = "red",
  alpha = 0.5,
  add = FALSE,
  ...
)
}
\arguments{
\item{f}{an expression defining the first function.}

\item{bottom}{defined the "bottom" of the shaded region. If you give a
number, then the bottom is just the number. You may also give an expression
defining a second function to shade the area between two curves.}

\item{xlim}{the region above which to draw}

\item{col}{the color of the shaded region}

\item{alpha}{the transparency of the shaded region}

\item{add}{whether to add the solid region to current figure or start anew}
}
\description{
Plots a filled region defined by one or two functions.
}
\examples{
f=mosaic::makeFun(sin(x)~x)
plotFunFill(f(x)~x,xlim=c(-3,3))
#mosaic::plotFun(f(x)~x,col="black",lwd=2,add=TRUE)
#mosaic::ladd(panel.grid(h=-1,v=-1,col="grey"))
f=mosaic::makeFun(sin(x)~x)
plotFunFill(f(x)~x,bottom=-2,xlim=c(-3,3))
#mosaic::plotFun(f(x)~x,col="black",lwd=2,add=TRUE)
#mosaic::ladd(panel.grid(h=-1,v=-1,col="grey"))
f=mosaic::makeFun(sin(x)~x)
g=mosaic::makeFun(-x-1/2*cos(5*x)~x)
plotFunFill(f(x)~x,bottom=g(x)~x,xlim=c(0,3))
#mosaic::plotFun(f(x)~x,col="black",lwd=2,add=TRUE)
#mosaic::plotFun(g(x)~x,col="black",lwd=2,add=TRUE)
#mosaic::ladd(panel.grid(h=-1,v=-1,col="grey"))

}
