% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVectorField.R
\name{plotODEDirectionField}
\alias{plotODEDirectionField}
\title{Plot a direction field for a scalar ODE}
\usage{
plotODEDirectionField(
  expression,
  tlim = c(0, 10),
  ylim = c(-5, 5),
  ics = NA,
  N = 20,
  ylist = NA,
  col = "black",
  lwd = 2,
  add = FALSE,
  plot = lattice::trellis.last.object(),
  ...
)
}
\arguments{
\item{expression}{an expression giving the right hand side of the ODE, as a function of t and y, in that order.}

\item{tlim}{the extent of the horizonatal axis}

\item{ylim}{the extend of the vertical axis}

\item{ics}{If desired, a list of initial values from which to draw trajectories.}

\item{N}{the number of horizontal and vertical rows of arrows to draw}

\item{ylist}{the vertical grid points to use in the plot, if specified.}

\item{col}{arrow color}

\item{lwd}{arrow line width}

\item{add}{T/F variable indicating whether this should be added to plot}

\item{plot}{the plot to which this should be added, default is the current plot.}

\item{...}{additional arguments to pass to xyplot arrows.}
}
\description{
Plot a direction field for a scalar ODE
}
\examples{
 #Consider the ODE
 #y'=1/2*y+cos(t)
 #Just give the plotter the right hand side with an expression as usual.  You may omit t
 #if the equation is autnomous. If you include t, it *must* be the first argument.
 plotODEDirectionField(1/2*y+cos(t)~t&y)



 #You can set the color, line width, and type as well.
 plotODEDirectionField(1/2*y+cos(t)~t&y,col="black",lwd=2)

 #You can add it to another plot, or another plot to it.
 results=Euler(1/2*y+cos(t)~t&y,tlim=c(0,10),ic=0,stepSize=0.1)
 mosaic::plotPoints(y~t,data=results,add=TRUE)

 #You can give an autonomous equation and not specify t as an input.
 plotODEDirectionField(-y*(y-1)~y,tlim=c(0,4),ylim=c(-0.1,2.1),N=8)

 #You can also just plot a vector field and add on trajectories from different initial
 #conditions using the y0s option. Add one initial value for every trajectory you want.
 plotODEDirectionField(1/2*y+cos(t)~t&y,ics=c(-2,0,-1,2))

}
