% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVectorField.R
\name{plotODEDirectionField}
\alias{plotODEDirectionField}
\title{Plot a direction field for a scalar ODE}
\usage{
plotODEDirectionField(
  expression,
  tlim = c(0, 10),
  ylim = c(-5, 5),
  y0s = NA,
  grid.by = (tlim[[2]] - tlim[[1]])/19,
  radius = grid.by * 0.8,
  col = NA,
  lty = 1,
  lwd = 1,
  add = FALSE,
  plot = lattice::trellis.last.object()
)
}
\arguments{
\item{expression}{an expression giving the right hand side of the ODE, as a function of t and y, in that order.}

\item{tlim}{the extent of the horizonatal axis}

\item{ylim}{the extend of the vertical axis}

\item{y0s}{If desired, a list of initial values from which to draw trajectories.}

\item{grid.by}{the spacing between arrow bases}

\item{radius}{the size of the arrow heads}

\item{col}{arrow color}

\item{lty}{arrow line type}

\item{lwd}{arrow line width}

\item{add}{T/F variable indicating whether this should be added to plot}

\item{plot}{the plot to which this should be added, default is the current plot.}
}
\description{
Plot a direction field for a scalar ODE
}
\examples{
 #Consider the ODE
 #y'=1/2*y+cos(t)
 #Just give the plotter the right hand side with an expression as usual.  The expression
 #MUST have both t and y as variables on the right hand side, even if it is autonomous
 #and only y shows up in the formula.
 plotODEDirectionField(1/2*y+cos(t)~t&y)

 #You can set the color, line width, and type as well.
 plotODEDirectionField(1/2*y+cos(t)~t&y,col="black",lwd=2)

 #You can add it to another plot, or another plot to it.
 results=Euler(1/2*y+cos(t)~t&y,tlim=c(0,10),y0=0,stepSize=0.1)
 mosaic::plotPoints(y~t,data=results,add=TRUE)

 #You can also just plot a vector field and add on trajectories from different initial
 #conditions using the y0s option. Add one initial value for every trajectory you want.
 plotODEDirectionField(1/2*y+cos(t)~t&y,y0s=c(-2,0,-1,2))

}
