% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVectorField.R
\name{plotVectorField}
\alias{plotVectorField}
\title{Plot a vector field
Scales all vectors to have same length}
\usage{
plotVectorField(
  expression,
  xlim = c(-5, 5),
  ylim = c(-5, 5),
  grid.by = (xlim[[2]] - xlim[[1]])/19,
  radius = grid.by * 0.8,
  col = NA,
  lwd = 1,
  lty = 1,
  add = FALSE,
  plot = lattice::trellis.last.object()
)
}
\arguments{
\item{expression}{A function of two variables that returns two ouputs to plot.}

\item{xlim}{the limits of the horizontal axis}

\item{ylim}{the limits of the vertical axis}

\item{grid.by}{the spacing between arrow bases}

\item{radius}{the size of the arrow heads}

\item{col}{arrow color}

\item{lwd}{arrow line width}

\item{lty}{arrow line type}

\item{add}{T/F variable indicating whether this should be added to plot}

\item{plot}{the plot to which this should be added, default is the current plot.}
}
\description{
Plot a vector field
Scales all vectors to have same length
}
\examples{
# example code
f=makeFun(c(-y,x)~x&y)
plotVectorField(f(x,y)~x&y, xlim=c(-3,3),ylim=c(-4,4))

#You can set the color, line width, or line type as well.
plotVectorField(f(x,y)~x&y, xlim=c(-3,3),ylim=c(-4,4),col="green",lwd=3,lty=1)

#You can add it to other plots if you so choose.
plotFun(sin(x)~x,xlim=c(0,5))
plotVectorField(f(x,y)~x&y, xlim=c(-3,3),ylim=c(-4,4),col="green",lwd=3,lty=1,add=TRUE)
}
